requirements.txt
# Screen & Image
mss>=9.0
opencv-python>=4.9
pyautogui>=0.9
Pillow>=10.0

# OCR
rapidocr-onnxruntime>=1.3

# LLM
openai>=1.30
jinja2>=3.1

# Data
aiosqlite>=0.20

# Notification
python-telegram-bot>=21.0
aiohttp>=3.9

# Base
pyyaml>=6.0
pydantic>=2.5